Added "outputLogs" Configuration Option (#1332)
* Added "outputLogs" To `turbo.json`

This commit adds a new configuration option
"outputLogs" which will be used to set the
default value of the `--output-logs` run option.

* Refactored `--output-logs` Run Option

This commit refactors the task log output to
use the new `util.TaskOutputMode`and
removes the `LogMode` type used previously.

* Added "outputLog" Tests

* Added "outputLogs" Documentation

* Simplified TaskOutputMode Constants

* Changed TaskOutputMode Into Enum

This commit replaces the TaskOutputMode strings
with an appropriate enum and related parsing. It
now uses UnmarshalJSON to parse as well and so
we can remove `ValidateTurboJSON` entirely.

* Moved "outputLogs" Config Into Pipeline

This commit moves the root "outputLogs"
configuration option into the pipeline config.
It also renames it to "outputMode" to
better reflect the nature of the option.

* Fixed Accidental Changes

* Removed More Unnecessary Changes

After scrolling through all of the changed files in GitHub,
I found other instances of unnecessary changes. Since
this PR went through a few iterations, some things were
left behind that weren't necessary.

* Added `"outputMode"` Parsing Test

* Refactored TaskOutputModeOverride Into TaskCache

This removes the need to deal with strings outside of
the "--output-logs" value handling.

* Added --output-logs Default Value Error Handling

* Mark TaskOutputModeOverride as optional in runcache.Opts. Remove flags plumbing from run

* Clarified `RunCache.taskOutputMode` As Override

Co-authored-by: Greg Soltis <greg.soltis@vercel.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>