docs: add N26 to showcase list (#1610)
Hey folks 👋 ,
First of all, thanks for the amazing work with Turborepo. I've started experimenting with this tool in my current work and it works so well that we managed to get it fully integrated into our pipelines.

This little PR is including https://n26.com to the showcase list.

| Light Version | Dark Version |
|------|-------|
| <img width="1840" alt="light" src="https://user-images.githubusercontent.com/2049560/182236501-9b9a6187-a616-45ed-87f0-c96fdd890fef.png">    | <img width="1840" alt="dark" src="https://user-images.githubusercontent.com/2049560/182236562-71bd10a5-b57d-4101-aa8a-046e9768ac35.png">    |

Cheers!

Co-authored-by: Jared Palmer <4060187+jaredpalmer@users.noreply.github.com>