Document inclusion of package.json in workspace task cache keys (#1955)
This PR documents #1832 now that we're releasing it.

**Edit: Do not merge this until 1.4.7 is released**

In doing so, I'm also changing the structure of the docs to separate env and files, which also needs to happen for #1936.

One intermediate unfortunate side effect is that the docs, for some time until #1936 is merged, will document `globalDependencies` separately in two sections. This may be a little confusing during this time frame.